refactor(schemas): migrate Phase 29 endpoint schemas — 2 files, 9 classes (#6042)
- knowledge_ai_stack.py: migrate EnhancedSearchRequest→AIStackEnhancedSearchRequest,
  KnowledgeExtractionRequest→AIStackKnowledgeExtractionRequest,
  DocumentAnalysisRequest, RAGQueryRequest→AIStackRAGQueryRequest
  → schemas_knowledge.py
- marketplace_sources.py: migrate MarketplaceSource, MarketplaceSourceCreate,
  MarketplaceSourcesResponse, CatalogPlugin, CatalogDocument
  → schemas_workflows.py (HttpUrl import added)

Renames to avoid collisions with EXISTING (different) classes in schemas_knowledge.py:
- EnhancedSearchRequest → AIStackEnhancedSearchRequest (existing has tag filtering, pagination)
- KnowledgeExtractionRequest → AIStackKnowledgeExtractionRequest (existing has fewer fields)
- RAGQueryRequest → AIStackRAGQueryRequest (existing has different fields)

The competing definitions are documented as a future de-duplication opportunity.
